refactor: Route channel lead nudges through Effect pipeline
- Add Effect::NudgeChannelLead { task_id, message } variant so channel
  lead nudges flow through execute_effects() like all other side effects
- Add channel_lead_sessions to WorldSnapshot for effect executor access
- Remove nudge_channel_lead_for_task() async helper from mod.rs (was
  bypassing the Effect pipeline with direct session_manager calls)
- Fix redundant extract_task_id_from_pr_title call in PR-opened handler
  by reusing the pr_task_id binding already computed above
- Move rpc_task.rs inline tests to separate rpc_task_tests.rs per CLAUDE.md
