+++ src/daemon/dispatch_tests.rs
@@ -990,6 +990,7 @@ fn test_spawn_for_pending_tasks_generates_registry_effects_new_task() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -1142,6 +1143,7 @@ fn test_spawn_for_pending_tasks_reuses_worktree_for_owned_task() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1264,6 +1266,7 @@ fn test_spawn_for_pending_tasks_skips_when_owner_has_pending_task() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1359,6 +1362,7 @@ fn test_spawn_owner_includes_record_task_assignment_for_cross_tick_dedup() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1458,6 +1462,7 @@ fn test_cross_tick_dedup_skips_in_flight_owned_task() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1573,6 +1578,7 @@ fn test_cross_case_dedup_prevents_same_coworker_from_case1_and_case2() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         tasks_with_worktrees: HashSet::new(),
         task_worktree_map: HashMap::new(),
         worktree_branch_owners: HashMap::new(),
@@ -1689,6 +1695,7 @@ fn test_spawn_for_pending_tasks_skips_via_snapshot_assignment_check() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1781,6 +1788,7 @@ fn test_orphan_recovery_reuses_existing_task_worktree() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -1927,6 +1935,7 @@ fn test_orphan_recovery_creates_new_worktree_when_none_exists() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -2093,6 +2102,7 @@ fn test_spawn_for_pending_unowned_reuses_existing_worktree() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -2237,6 +2247,7 @@ fn make_reconcile_snapshot(
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -2751,6 +2762,7 @@ fn test_spawn_for_pending_tasks_when_all_coworkers_are_gone() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -3407,6 +3419,7 @@ fn test_spawn_extracts_model_alias_from_provider_model_format() {
         task_model_map,
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -3506,6 +3519,7 @@ fn test_orphan_recovery_marks_task_in_flight() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         merged_pr_numbers: HashSet::new(),
@@ -3785,6 +3799,7 @@ fn test_stale_task_cleanup_false_positive_task_about_merged_pr() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
@@ -3884,6 +3899,7 @@ fn test_stale_task_cleanup_correct_behavior_with_explicit_pr_field() {
         task_model_map: HashMap::new(),
         task_plan_map: HashMap::new(),
         task_execution_skill_map: HashMap::new(),
+        channel_lead_sessions: HashMap::new(),
         coworkers_with_open_prs: HashSet::new(),
         coworkers_with_merged_prs: HashSet::new(),
         ci_passed_pr_coworkers: HashSet::new(),
